requirements.txt
torch
ultralytics
opencv-python
numpy
matplotlib
pandas
filterpy
lap
scikit-learn
pyyaml
